Move tests for receiving completedblock messages.
